pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maunium-stickerpicker"
version = "0.2.0"
description = "A fast and simple Matrix sticker picker widget"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "GNU Affero General Public License v3 or later (AGPLv3+)" }
authors = [{ name = "Tulir Asokan", email = "tulir@maunium.net" }]
urls = { "Homepage" = "https://github.com/maunium/stickerpicker" }
classifiers = [
	"Development Status :: 4 - Beta",
	"License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
	"Framework :: AsyncIO",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
]
dependencies = [
	"aiohttp",
	"yarl",
	"pillow",
	"telethon",
	"cryptg",
	"python-magic",
	"rlottie-python",
]

[project.optional-dependencies]
test = ["pytest>=8.3.0", "pytest-asyncio>=0.21.0", "pytest-mock>=3.10.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[project.scripts]
sticker-import = "sticker.stickerimport:cmd"
sticker-pack = "sticker.pack:cmd"

[tool.setuptools]
package-dir = { "" = "src" }
